Split login strategies into separate, reusable classes
